refactor(counter): moved `executeCounter` into its own file and additionally export `CounterModule` as `Counter`
