+++ packages/counter/src/CounterModule.ts
@@ -1,6 +1,7 @@
 import type { CounterReport } from '@qualweb/core/evaluation';
 import { ExecutableModuleContext, ModuleType } from '@qualweb/core/evaluation';
 import type { QualwebPage } from '@qualweb/core/lib';
+import { executeCounter } from './executeCounter';
 
 export class CounterModule extends ExecutableModuleContext {
   public readonly name = ModuleType.COUNTER;
@@ -11,8 +12,7 @@ export class CounterModule extends ExecutableModuleContext {
 
   protected runModule(page: QualwebPage): Promise<CounterReport> {
     return page.evaluate(() => {
-      //@ts-expect-error The package exists inside the context of the WebPage
-      return window.executeCounter();
+      return executeCounter();
     });
   }
 }
